+++ package.json
@@ -1,63 +1,63 @@
 {
-	"name": "tvoi-ff",
-	"private": true,
-	"version": "0.0.0",
-	"type": "module",
-	"scripts": {
-		"dev": "vite",
-		"build": "vite build",
-		"build:development": "vite build --mode development",
-		"build:testing": "vite build --mode testing",
-		"preview": "vite preview",
-		"test": "vitest",
-		"lint": "eslint --ext .ts,.vue .",
-		"lint:fix": "eslint --ext .ts,.vue . --fix",
-		"prettier": "prettier . --check",
-		"prettier:write": "prettier . --write",
-		"stylelint": "stylelint \"**/*.{vue,css}\"",
-		"stylelint:fix": "stylelint \"**/*.{vue,css}\" --fix",
-		"format": "bun run lint:fix && bun run prettier:write && bun run stylelint:fix",
-		"check": "bun run lint && bun run prettier && bun run stylelint"
-	},
-	"dependencies": {
-		"axios": "^1.6.7",
-		"markdown-it": "^14.0.0",
-		"pinia": "^2.1.7",
-		"query-string": "^9.0.0",
-		"ua-parser-js": "^1.0.37",
-		"vue": "^3.4.21",
-		"vue-material-design-icons": "^5.3.0",
-		"vue-router": "^4.3.0",
-		"vuetify": "^3.5.7",
-		"workbox-window": "^7.0.0"
-	},
-	"devDependencies": {
-		"@types/markdown-it": "^13.0.7",
-		"@types/node": "^20.11.24",
-		"@types/ua-parser-js": "^0.7.39",
-		"@typescript-eslint/eslint-plugin": "^7.1.0",
-		"@typescript-eslint/parser": "^7.1.0",
-		"@vitejs/plugin-basic-ssl": "^1.1.0",
-		"@vitejs/plugin-vue": "^5.0.4",
-		"@vue/eslint-config-typescript": "^12.0.0",
-		"eslint": "^8.57.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-prettier": "^5.1.3",
-		"eslint-plugin-vue": "^9.22.0",
-		"postcss": "^8.4.35",
-		"postcss-html": "^1.6.0",
-		"postcss-preset-env": "^9.4.0",
-		"prettier": "^3.2.5",
-		"stylelint": "^16.2.1",
-		"stylelint-config-recommended-vue": "^1.5.0",
-		"stylelint-config-standard": "^36.0.0",
-		"typescript": "5.3.3",
-		"vite": "^5.1.4",
-		"vite-plugin-eslint": "^1.8.1",
-		"vite-plugin-pwa": "^0.19.2",
-		"vite-plugin-stylelint": "^5.3.1",
-		"vite-plugin-vuetify": "^2.0.2",
-		"vitest": "^1.3.1",
-		"vue-tsc": "^2.0.2"
-	}
+  "name": "tvoi-ff",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "build:development": "vite build --mode development",
+    "build:testing": "vite build --mode testing",
+    "preview": "vite preview --port 443",
+    "test": "vitest",
+    "lint": "eslint --ext .ts,.vue .",
+    "lint:fix": "eslint --ext .ts,.vue . --fix",
+    "prettier": "prettier . --check",
+    "prettier:write": "prettier . --write",
+    "stylelint": "stylelint \"**/*.{vue,css}\"",
+    "stylelint:fix": "stylelint \"**/*.{vue,css}\" --fix",
+    "format": "bun run lint:fix && bun run prettier:write && bun run stylelint:fix",
+    "check": "vue-tsc && bun run lint && bun run prettier && bun run stylelint"
+  },
+  "dependencies": {
+    "axios": "^1.6.7",
+    "markdown-it": "^14.0.0",
+    "pinia": "^2.1.7",
+    "query-string": "^9.0.0",
+    "ua-parser-js": "^1.0.37",
+    "vue": "^3.4.21",
+    "vue-material-design-icons": "^5.3.0",
+    "vue-router": "^4.3.0",
+    "vuetify": "^3.5.7",
+    "workbox-window": "^7.0.0"
+  },
+  "devDependencies": {
+    "@types/markdown-it": "^13.0.7",
+    "@types/node": "^20.11.24",
+    "@types/ua-parser-js": "^0.7.39",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
+    "@vitejs/plugin-basic-ssl": "^1.1.0",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-vue": "^9.22.0",
+    "postcss": "^8.4.35",
+    "postcss-html": "^1.6.0",
+    "postcss-preset-env": "^9.4.0",
+    "prettier": "^3.2.5",
+    "stylelint": "^16.2.1",
+    "stylelint-config-recommended-vue": "^1.5.0",
+    "stylelint-config-standard": "^36.0.0",
+    "typescript": "5.3.3",
+    "vite": "^5.1.4",
+    "vite-plugin-eslint": "^1.8.1",
+    "vite-plugin-pwa": "^0.19.2",
+    "vite-plugin-stylelint": "^5.3.1",
+    "vite-plugin-vuetify": "^2.0.2",
+    "vitest": "^1.3.1",
+    "vue-tsc": "^2.0.2"
+  }
 }
